Install eslint-plugin-react-refresh
