+++ client/CLAUDE.md
@@ -15,9 +15,9 @@ Note: `find-unused-translation-keys` is a **server** script (it scans `client/sr
 ## Directory Layout (`client/src`)
 
 - **`index.tsx`** — entry point: Sentry init, Redux `Provider`, styled-components `ThemeProvider`, then the lazily-loaded `App`.
-- **`app/`** — `App.tsx` (initial data load + router wrapper), `AppRoutes.tsx` (all react-router routes), `lazyViews.ts` (the code-split views and their prefetch), `HistoryContext.tsx` (tracks previous location for back buttons).
+- **`app/`** — `App.tsx` (initial data load + router wrapper), `AppRoutes.tsx` (all react-router routes), `lazyViews.ts` (the code-split views and their prefetch), `HistoryProvider.tsx` + `historyContext.ts` (tracks previous location for back buttons).
 - **`views/`** — one folder per page/feature: `about`, `admin`, `admission-ticket`, `all-program-items`, `dashboard`, `event-log`, `group`, `helper`, `login`, `logout`, `my-program-items`, `profile`, `program-item`, `registration`. A view folder typically holds `*View.tsx`, its Redux `*Slice.ts`, `*Thunks.ts`, and a local `components/`.
-- **`components/`** — shared UI primitives (Button, Input, Checkbox, Header, etc.) and `icons/` (FontAwesome).
+- **`components/`** — shared UI primitives (Button, Input, Checkbox, Header, etc.) and `icons/` (FontAwesome). `componentStyles.ts` collects the style variant enums those primitives take as props (`ButtonStyle`, `HighlightStyle`, `InfoTextVariant`) — kept out of the `.tsx` files so Fast Refresh keeps working (see Build & Environment).
 - **`services/`** — typed API call wrappers, one file per domain (`loginServices.ts`, `programItemsServices.ts`, `groupServices.ts`, …). Each just wraps `api.post/get/delete` with `shared/types/api/...` types.
 - **`state/`** — only `loading/loadingSlice.ts` lives here; **the store itself and most slices do not** (see State Management below).
 - **`locales/`** — `en.ts` and `fi.ts` translation objects.
@@ -77,6 +77,7 @@ i18next, English + Finnish. Locale files: `client/src/locales/{en,fi}.ts` (deepl
 ## Build & Environment
 
 - **Vite** (`client/vite.config.ts`): outputs to `client/build/` with sourcemaps; build target derived from `.browserslistrc`; plugins for MDX, SVGR (SVG→React), React, static copy, and gzip/brotli compression; `tsconfigPaths` resolves the `client`/`shared` aliases; dev server bound to `127.0.0.1:8000`.
+- **Fast Refresh is lint-enforced:** `eslint-plugin-react-refresh`'s `vite` config runs over `client/**` in the root `eslint.config.ts`. A module that exports a component plus anything else loses Fast Refresh and falls back to a full page reload, so enums, hooks, helpers and standalone styled-components shared between modules live in their own file (`components/componentStyles.ts`, `app/historyContext.ts`, …) rather than beside the component that first needed them. The `vite` variant allows constant exports, which `@vitejs/plugin-react` handles.
 - **Chunking** (`build.rolldownOptions.output.codeSplitting`): dependencies are grouped out of the app's own chunks into `react` (react/react-dom/scheduler), `sentry` (`@sentry/*`) and `vendor` (the rest of `node_modules`), so their content hashes survive an app-only deploy and a returning client re-downloads app code alone. Use `codeSplitting` — `advancedChunks` is the deprecated spelling of the same option. Group `test` regexes match module ids, which are Windows paths locally, so write separators as `[\\/]`.
 - **Boot preloads** (`client/preloadBootChunksPlugin.ts`): Vite emits `<link rel="modulepreload">` only for the entry's **static** import graph, so chunks behind a dynamic import don't start downloading until the entry has downloaded, instantiated its dependencies and run. The plugin names modules that are needed on every load anyway (currently `app/App.tsx`) and preloads their chunks, minus whatever Vite already covers. It fails the build if a named module no longer resolves to a chunk, so a rename can't silently drop the hint. Only list always-needed modules here — a genuinely deferred view should keep discovering its chunk at navigation time.
 - **React Compiler is ON**, wired as `@rolldown/plugin-babel` running the plugin's `reactCompilerPreset`. It has to go through Babel: `@vitejs/plugin-react` v6 transforms JSX with Oxc and takes **no** `babel` option of its own, so the `babel: { plugins: [...] }` that used to enable the compiler was silently ignored from the Vite 8 upgrade onwards and the codebase spent months un-memoized. The coverage build (`COVERAGE=true`) leaves it out: the compiler's rewrites move the statements istanbul instrumented out from under the positions it recorded. `babel-plugin-react-compiler` pins `@babel/core` to 7.x — it depends on `@babel/types@^7`, and Babel 8 would give it a second copy.
@@ -84,7 +85,7 @@ i18next, English + Finnish. Locale files: `client/src/locales/{en,fi}.ts` (deepl
   Enabling it meant fixing what it exposed, none of which `yarn lint`, `yarn type-check` or the unit suite could see:
   - **`useTimeNow`** used to call `dayjs()` while rendering. The compiler had nothing to key that on, so it cached the instant behind the first-render sentinel and froze the clock at whatever each component mounted with. The mocked-time branch keys on `testTime` and stayed correct, so the **entire Playwright suite passed** while real users saw a frozen clock. It now reads a shared ticking store (`utils/clock.ts`) through `useSyncExternalStore`, which also makes the value referentially stable between ticks, so it can be listed as a dependency like anything else. That store is unit-tested (`utils/clock.test.ts`) precisely because the E2E suite runs on mocked time and cannot see this class of bug.
   - **`AllProgramItemsList`** is the one deliberate opt-out (`"use no memo"`): `useWindowVirtualizer` returns a render-stable instance whose measurements mutate inside it, so the row offsets and total size read from it during render are external mutable state. Memoized, rows keep the offsets they first rendered with and a sign-up form's Confirm button ends up underneath the next row (caught by `lotterySignup` specs 209/300).
-  - **A ref read during render** makes the compiler decline the component outright — `HistoryContext` and `EventLog` both did. Both hold the value in state now; in `HistoryContext` that also fixed a real bug, where the current location promoted itself to being its own predecessor on any re-render that wasn't a navigation, while `EventLog`'s mount snapshot behaves the same either way.
+  - **A ref read during render** makes the compiler decline the component outright — `HistoryProvider` and `EventLog` both did. Both hold the value in state now; in `HistoryProvider` that also fixed a real bug, where the current location promoted itself to being its own predecessor on any re-render that wasn't a navigation, while `EventLog`'s mount snapshot behaves the same either way.
   - **Only a plain `try`/`catch` is supported.** A `finally` is declined (`Handle TryStatement with a finalizer`) and so is a catch-less `try`, and either one declines the entire enclosing function. The polling loop needs its `finally`, so it moved out of `App.tsx` into `utils/dataPolling.ts`, where the compiler doesn't look; the Kompassi redirect handler dropped its `finally` for a `catch` plus a reset after the block, and stayed where it was.
   - **Mutating a `useState` array in place** and passing the same reference back (`AdminActionCard`) — an immutable update now.
