+++ client/src/components/ErrorBar.tsx
@@ -1,7 +1,7 @@
 import { ReactElement } from "react";
 import { useTranslation } from "react-i18next";
 import { DismissibleBanner } from "client/components/DismissibleBanner";
-import { HighlightStyle } from "client/components/RaisedCard";
+import { HighlightStyle } from "client/components/componentStyles";
 import { BackendErrorType } from "client/types/errorTypes";
 import { useAppDispatch, useAppSelector } from "client/utils/hooks";
 import { removeError } from "client/views/admin/adminSlice";
