+++ client/src/views/event-log/NotificationBar.tsx
@@ -5,7 +5,8 @@ import styled from "styled-components";
 import { config } from "shared/config";
 import { AboutTab } from "client/app/routes";
 import { DismissibleBanner } from "client/components/DismissibleBanner";
-import { HighlightStyle, RaisedCard } from "client/components/RaisedCard";
+import { RaisedCard } from "client/components/RaisedCard";
+import { HighlightStyle } from "client/components/componentStyles";
 import { useAppDispatch, useAppSelector } from "client/utils/hooks";
 import { EventLogEventMessage } from "client/views/event-log/EventLogEventMessage";
 import { submitUpdateEventLogIsSeen } from "client/views/login/loginThunks";
