+++ client/src/views/group/components/GroupCreatorActions.tsx
@@ -1,8 +1,9 @@
 import { ReactElement, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { Button, ButtonStyle } from "client/components/Button";
+import { Button } from "client/components/Button";
 import { ButtonGroup } from "client/components/ButtonGroup";
 import { ErrorMessage } from "client/components/ErrorMessage";
+import { ButtonStyle } from "client/components/componentStyles";
 import { useAppDispatch } from "client/utils/hooks";
 import {
   PostCloseGroupErrorMessage,
