+++ client/src/views/group/components/GroupMemberActions.tsx
@@ -1,7 +1,8 @@
 import { ReactElement, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { Button, ButtonStyle } from "client/components/Button";
+import { Button } from "client/components/Button";
 import { ErrorMessage } from "client/components/ErrorMessage";
+import { ButtonStyle } from "client/components/componentStyles";
 import { useAppDispatch } from "client/utils/hooks";
 import {
   PostLeaveGroupErrorMessage,
