+++ client/src/views/program-item/body/components/ProgramItemDetails.tsx
@@ -4,7 +4,8 @@ import { useTranslation } from "react-i18next";
 import { capitalize } from "remeda";
 import styled from "styled-components";
 import { ProgramItem } from "shared/types/models/programItem";
-import { InfoText, InfoTextVariant } from "client/components/InfoText";
+import { InfoText } from "client/components/InfoText";
+import { InfoTextVariant } from "client/components/componentStyles";
 import { TextWithLinks } from "client/markdown/components/TextWithLinks";
 import { useTimeNow } from "client/utils/useTimeNow";
 import { getFormattedInterval } from "client/views/program-item/programItemUtils";
