+++ client/src/views/program-item/signup/components/direct-signup/ProgramItemDirectSignup.tsx
@@ -3,10 +3,10 @@ import { useTranslation } from "react-i18next";
 import styled from "styled-components";
 import { ProgramItem } from "shared/types/models/programItem";
 import { getDirectSignupStartTime } from "shared/utils/signupTimes";
-import { ButtonStyle } from "client/components/Button";
 import { ErrorMessage } from "client/components/ErrorMessage";
 import { InfoText } from "client/components/InfoText";
 import { SignupQuestionAnswer } from "client/components/SignUpQuestionAnswer";
+import { ButtonStyle } from "client/components/componentStyles";
 import { startLoading, stopLoading } from "client/state/loading/loadingSlice";
 import { useAppDispatch, useAppSelector } from "client/utils/hooks";
 import { useTimeNow } from "client/utils/useTimeNow";
