+++ client/src/views/program-item/signup/components/lottery-signup/ProgramItemLotterySignup.tsx
@@ -2,9 +2,9 @@ import { ReactElement, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { ProgramItem } from "shared/types/models/programItem";
 import { getLotterySignupStartTime } from "shared/utils/signupTimes";
-import { ButtonStyle } from "client/components/Button";
 import { ErrorMessage } from "client/components/ErrorMessage";
 import { InfoText } from "client/components/InfoText";
+import { ButtonStyle } from "client/components/componentStyles";
 import { startLoading, stopLoading } from "client/state/loading/loadingSlice";
 import { useAppDispatch, useAppSelector } from "client/utils/hooks";
 import { useTimeNow } from "client/utils/useTimeNow";
