+++ package.json
@@ -100,6 +100,7 @@
     "eslint-plugin-react": "7.37.5",
     "eslint-plugin-react-hooks": "7.1.1",
     "eslint-plugin-react-hooks-addons": "0.5.1",
+    "eslint-plugin-react-refresh": "0.5.4",
     "eslint-plugin-unicorn": "67.0.0",
     "globals": "17.6.0",
     "istanbul-lib-coverage": "3.2.2",
